# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "auth_register",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "auth_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "auth_refresh",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/device/approve": {
      "post": {
        "operationId": "auth_approveDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подтверждение кода из расширения — юзер уже авторизован в вебе",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mobile/register": {
      "post": {
        "operationId": "authMobile_register",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth-mobile"
        ]
      }
    },
    "/auth/mobile/login": {
      "post": {
        "operationId": "authMobile_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth-mobile"
        ]
      }
    },
    "/auth/mobile/refresh": {
      "post": {
        "operationId": "authMobile_refresh",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileAuthResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth-mobile"
        ]
      }
    },
    "/auth/mobile/logout": {
      "post": {
        "operationId": "authMobile_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "auth-mobile"
        ]
      }
    },
    "/auth/mobile/device/code": {
      "post": {
        "operationId": "authMobile_createDeviceCode",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodeResponse"
                }
              }
            }
          }
        },
        "summary": "Расширение просит код и показывает его пользователю",
        "tags": [
          "auth-mobile"
        ]
      }
    },
    "/auth/mobile/device/poll": {
      "post": {
        "operationId": "authMobile_pollDeviceCode",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePollResponse"
                }
              }
            }
          }
        },
        "summary": "Расширение опрашивает эндпоинт, пока юзер не подтвердит код",
        "tags": [
          "auth-mobile"
        ]
      }
    },
...
    "/ai/recommendations": {
      "get": {
        "operationId": "ai_getRecommendations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSuggestionsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Рекомендации по вкусовому профилю",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/picker": {
      "post": {
        "operationId": "ai_pick",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSuggestionsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "AI Picker — подбор под настроение",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/summary/{slug}": {
      "get": {
        "operationId": "ai_getSummary",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Описание без спойлеров для детальной страницы",
        "tags": [
          "ai"
        ]
      }
    }
  },
  "info": {
    "title": "RED Marathōn API",
    "description": "One library for movies, TV shows, anime, books and games",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 100,
            "format": "email"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "pattern": "^[a-zA-Z0-9_]+$"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "AuthUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MODERATOR",
              "ADMIN"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponse"
          },
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 100,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "DeviceCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{3}-\\d{3}$"
          }
        },
        "required": [
          "code"
        ]
      },
      "MobileAuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponse"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "DeviceCodeResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Шесть цифр в формате 123-456"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "intervalSec": {
            "type": "number",
            "description": "Через сколько секунд опрашивать /auth/device/poll"
          }
        },
        "required": [
          "code",
          "expiresAt",
          "intervalSec"
        ]
      },
      "DevicePollResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ]
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "TitleListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "coverUrl",
          "releaseDate",
          "rating",
          "ratingCount"
        ]
      },
...
      "AiSuggestionResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ]
          },
          "reason": {
            "type": "string",
            "description": "Почему модель это предложила"
          },
          "title": {
            "nullable": true,
            "description": "Заполнен, если тайтл уже есть в нашей базе",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TitleListItemResponse"
              }
            ]
          }
        },
        "required": [
          "name",
          "type",
          "reason",
          "title"
        ]
      },
      "AiSuggestionsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiSuggestionResponse"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "PickerDto": {
        "type": "object",
        "properties": {
          "mood": {
            "type": "string",
            "maxLength": 200
          },
          "type": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GAME",
                "MOVIE",
                "TV_SHOW",
                "BOOK",
                "ANIME"
              ]
            }
          },
          "duration": {
            "type": "string",
            "enum": [
              "short",
              "medium",
              "long"
            ]
          },
          "excludeLibrary": {
            "type": "boolean",
            "description": "Не предлагать то, что уже есть в библиотеке",
            "default": true
          }
        }
      },
      "AiSummaryResponse": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "summary"
        ]
      }
    }
  }
}